+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+##
+## Changed
+- fix author in group folder
+
 ## 7.9.4
 ## Changed
 - remove link to docs cloud